requirements.txt
asgiref==3.5.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-autoslug==1.9.8
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
